Break Select sub-components out into separate files
