+++ packages/components/src/Form/Inputs/InputText/InputText.tsx
@@ -105,5 +105,3 @@ InputText.defaultProps = {
   height: CustomizableInputTextAttributes.height,
   type: 'text',
 }
-
-const foo = <InputText value="hey" />
