+++ packages/components/src/Form/Inputs/Select/index.ts
@@ -26,4 +26,6 @@
 
 export * from './Select'
 export * from './SelectContext'
-// export * from './SelectOption'
+export * from './SelectInput'
+export * from './SelectList'
+export * from './SelectOption'
