+++ packages/playground/src/index.tsx
@@ -24,7 +24,7 @@
 
  */
 
-import React, { useState } from 'react'
+import React from 'react'
 import ReactDOM from 'react-dom'
 import {
   GlobalStyle,
